feat: enhance Header and Home page styles, add back navigation button in Header
