+++ src/pages/Home/Home.css
@@ -1,5 +1,6 @@
 .home-row {
-  min-height: calc(100vh - 110px);
+  min-height: auto;
+  margin-bottom: 40px;
 }
 
 .carousel-content {
@@ -27,3 +28,9 @@
   box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);
 }
 
+.pokemon-list-row {
+  padding: 20px;
+}
+.pokemon-list-image {
+  width: inherit;
+}
