+++ src/pages/PokemonDetails/PokemonDetails.css
@@ -2,11 +2,8 @@
   padding: 20px;
   max-width: 1200px;
   margin: 0 auto;
-}
-
-.back-button {
-  margin-bottom: 20px;
-  font-size: 16px;
+  position: relative;
+  z-index: 1;
 }
 
 .pokemon-details-loading,
